Learn Building Blocks of Image Processing/src/python/requirements.txt
numpy==1.19.1
Pillow==7.2.0